calcular renta y listado de todos los clientes

+++ src/main/java/com/example/lab6_gtics_20221_g4/entity/Employee.java
@@ -35,7 +35,7 @@ public class Employee {
     private Job job;
 
     @Column(name = "salary", precision = 8, scale = 2)
-    private BigDecimal salary;
+    private float salary;
 
     @Column(name = "commission_pct", precision = 2, scale = 2)
     private BigDecimal commissionPct;
@@ -83,11 +83,11 @@ public class Employee {
         this.commissionPct = commissionPct;
     }
 
-    public BigDecimal getSalary() {
+    public float getSalary() {
         return salary;
     }
 
-    public void setSalary(BigDecimal salary) {
+    public void setSalary(float salary) {
         this.salary = salary;
     }
